Corrected an error with teh PRINT_ERROR and PRINT_WARNING macros and their use in the Flag's constructors. Fixed multi define issue from parse functions
